backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
bidict==0.23.1
black==25.1.0
boto3==1.38.18
botocore==1.38.18
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.4
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.111.0
fastapi-cli==0.0.7
filelock==3.18.0
flake8==7.2.0
frozenlist==1.7.0
fsspec==2025.3.2
h11==0.16.0
hf-xet==1.1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.31.2
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jmespath==1.0.1
jq==1.8.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
multidict==6.5.0
mypy==1.15.0
mypy_extensions==1.1.0
numpy==2.2.6
oauthlib==3.2.2
orjson==3.10.18
packaging==25.0
pandas==2.2.3
passlib==1.7.4
passlib[bcrypt]==1.7.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
pyasn1==0.6.1
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.9.1
pydantic_core==2.18.4
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.5.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.2
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rich-toolkit==0.14.8
rsa==4.9.1
s3transfer==0.12.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.37.2
tenacity==9.1.2
tomli==2.2.1
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.30.1
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
ddgs
slowapi==0.1.7
